Rename 'sources' directory to 'inputs'
