+++ internal-packages/workflow-designer-ui/src/editor/properties-panel/text-generation-node-properties-panel/input-panel.tsx
@@ -29,7 +29,7 @@ import {
 	type Source,
 	useConnectedSources,
 	useSourceCategories,
-} from "./sources";
+} from "./inputs";
 
 function SourceToggleItem({
 	source,
